refactor(comments): mv comment styles to right place
