+++ src/containers/unit/Comments/Comment/UpInfo.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { ICON_CMD } from '@/config'
 import { prettyNum } from '@/utils'
 
-import { Wrapper, Icon } from './styles/up_info'
+import { Wrapper, Icon } from '../styles/comment/up_info'
 
 import { toggleLikeComment } from '../logic'
